requirements.txt
streamlit==1.35.0
rank_bm25==0.2.2
nltk==3.8.1
gensim==4.3.2
numpy==1.26.4
requests==2.31.0
tqdm==4.66.2
scikit-learn==1.4.2
scipy==1.12
transformers==4.40.0
faiss-cpu==1.8.0  # 或faiss-gpu（如有CUDA）
torch==2.2.1
sentence-transformers==2.7.0

